backend/requirements.txt
django-ninja==0.18.0
PyMySQL==1.0.2
seldom~=2.10.7
Django~=4.1